+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### 29.04.2017
 
+* Rendering - added line numbers
 * Fixed cursor movement when merging lines by backspace (tab-related)
 * Added color to tabs in debug mode
 * Added TASKS.md